state: Move file review action scope
File review persistence and action-scope validation both live in
data.file_review.state, while records, freshness, commands, and selection
validation already have separate modules.

The state module therefore mixes disk-backed review data with
command-facing decisions about pathless actions, batch and live source
guards, and follow-up clearing. That makes the data package harder to
navigate because the module name points at persistence but exports action
resolution too.

This commit addresses that by adding data.file_review.action_scope for
review-scoped action validation and retargeting command callers to it,
leaving state.py with persisted review state IO.

The following tests commit will retarget tests and architecture guards to
the new action-scope module.
